Update readme.md
Fixed camel case for Elasticsearch, replaced Elastic with Elasticsearch (Elastic is the company, Elasticsearch is the software), removed spaces before colons, removed "the" before Data Collector, source => origin, capitalization and periods for consistency. 
Please make sure my changes didn't break links to files/video? 
Left "lat/lon" attributes in case that's the name of the attributes. But if it's not, then it's nicer to say "latitude and longitude attributes" - ?