Update log_shipping_to_elasticsearch_part1.md
Misc cleanup. A few things to try to remember: set up, fire up -- both two words. Data Collector is the product - when used alone, it shouldn't have a "the" in front of it. And I think property names, and tabs, stuff like that just needs to be capitalized, no quotation marks necessary. :)